apps/backend/services/liveportrait-worker/requirements.txt
# RabbitMQ
pika>=1.3.0

# Database
pymysql>=1.1.0
sqlalchemy>=2.0.0

# AWS S3
boto3>=1.34.0

# HTTP requests (for file download)
requests>=2.31.0

# Environment variables
python-dotenv>=1.0.0

# LivePortrait dependencies will be installed via conda
# torch, torchvision, etc. are managed by LivePortrait's environment
